+++ wattpad/src/main/java/com/iremayvaz/moderation/model/dto/ModerationResult.java
@@ -3,6 +3,7 @@ package com.iremayvaz.moderation.model.dto;
 import com.iremayvaz.moderation.model.enums.Severity;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @AllArgsConstructor
@@ -15,7 +16,7 @@ public class ModerationResult {
     private List<LabelPrediction> predictions;
 
     // Final ceza puanı
-    private Double score;
+    private BigDecimal score;
 
     // ML modeli hakkında debug/metadata (opsiyonel)
     private Severity severity;
